Personalise blog template with my details/images
Amend the _config.yml to have my personal details customising the
template to be my blog. Add images for the background and images
for my author persona.

Amend the README.md to describe the repository and guide users to the
template site if they wish to use this as their blog/site template.

Closes issue #1